requirements.txt
apiview==1.3.25
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.8
django-cors-headers==3.4.0
django-filter==2.3.0
django-filters==0.2.1
django-import-export==2.3.0
django-paypal==1.0.0
django-templated-mail==1.1.1
django-webpack-loader==0.7.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.4
paypal==1.2.5
Pillow==7.2.0
PyJWT==1.7.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
simplejwt==2.0.1
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
typing==3.7.4.3
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0
